pyproject.toml
[tool.poetry]
name = "wacz_signing"
version = "0.2.4"
description = "A library for signing and timestamping file hashes"
authors = ["Ben Steinberg <bsteinberg@law.harvard.edu>"]
readme = "README.md"
license = "GPL-3.0-or-later"

[tool.poetry.dependencies]
python = "^3.9"
cryptography = "^38.0.1"
pem = "^21.2.0"
rfc3161ng = "^2.1.3"
pyasn1 = "^0.4.8"
python-dotenv = "^0.21.0"


[tool.poetry.group.dev.dependencies]
flake8 = "^5.0.4"
pytest = "^7.1.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
